CFN CLI unit tests
