+++ lib/inspec-iggy/cloudformation/parser.rb
